new/requirements.txt
fastapi==0.115.5
motor==3.6.0
uvicorn
jinja2
python-multipart
openai==0.28
python-dotenv
paho-mqtt
bcrypt
thread